Buddy allocator before testing

+++ include/init.h
@@ -2,6 +2,6 @@
 #ifndef _INIT_H_
 #define _INIT_H_
 
-bool is_power_of_two(int);
-unsigned next_power_of_two(unsigned int);
+bool is_power_of_two(unsigned long);
+unsigned long next_power_of_two(unsigned long);
 #endif
