+++ memory.cpp
@@ -1,13 +1,13 @@
 //File: memory.cpp
 #include <init.h>
 
-bool is_power_of_two(unsigned int x){
+bool is_power_of_two(unsigned long x){
   return ((x != 0) && !(x & (x - 1)));
 }
 
-unsigned int next_power_of_two(unsigned int n)
+unsigned long next_power_of_two(unsigned long n)
 {
-    unsigned int p = 1;
+    unsigned long p = 1;
     if (is_power_of_two(n))
         return n;
